refactored dedicated ensemble utils
