+++ mlens/utils/__init__.py
@@ -7,8 +7,6 @@ author: Sebastian Flenenrhag
 date: 10/01/2017
 """
 
-from .utils import _slice, name_columns
-from .utils import pickle_save, pickle_load, print_time
+from .utils import _slice, pickle_save, pickle_load, print_time
 
-__all__ = ['_slice', 'name_columns', 'pickle_save',
-           'pickle_load', 'print_time']
+__all__ = ['_slice', 'pickle_save', 'pickle_load', 'print_time']
