+++ mlens/utils/utils.py
@@ -32,13 +32,6 @@ def _slice(Z, idx):
     return z
 
 
-def name_columns(estimator_cases):
-    """Utility func for naming a mapping of estimators on different cases"""
-    return [case + '-' + est_name if case not in [None, ''] else est_name
-            for case, estimators in estimator_cases.items()
-            for est_name, _ in estimators]
-
-
 def pickle_save(obj, name):
     """Utility function for pickling an object"""
     with open(name + '.pkl', 'wb') as f:
